requirements.txt
# CoC 跑团助手 - 依赖
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
openai>=1.12.0
pydantic>=2.5.0

# PDF 解析、RAG 检索、人物卡导入
pypdf>=4.0.0
pdfplumber>=0.11.0
rank_bm25>=0.2.2
jieba>=0.42.1
python-docx>=1.0.0
